# personal-project/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Movies API",
      "description": "API for managing a movie watchlist",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Local server"
      }
    ],
    "paths": {
      "/movies": {
        "get": {
          "summary": "Retrieve all movies",
          "tags": ["Movies"],
          "responses": {
            "200": {
              "description": "A list of movies"
            }
          }
        },
        "post": {
          "summary": "Create a new movie",
          "tags": ["Movies"],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "director": { "type": "string" },
                    "genre": { "type": "string" },
                    "releaseYear": { "type": "number" },
                    "rating": { "type": "number" },
                    "watched": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Movie created successfully"
            },
            "400": {
              "description": "Missing required fields"
            }
          }
        }
      },
      "/movies/{id}": {
        "get": {
          "summary": "Retrieve a movie by ID",
          "tags": ["Movies"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": { "type": "string" },
              "description": "The movie ID"
            }
          ],
          "responses": {
            "200": {
              "description": "A movie object"
            },
            "400": {
              "description": "Invalid ID format"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        },
        "put": {
          "summary": "Update an existing movie",
          "tags": ["Movies"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": { "type": "string" },
              "description": "The movie ID"
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "director": { "type": "string" },
                    "genre": { "type": "string" },
                    "releaseYear": { "type": "number" },
                    "rating": { "type": "number" },
                    "watched": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "responses": {
            "204": {
              "description": "Movie updated successfully"
            },
            "400": {
              "description": "Invalid ID format"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a movie",
          "tags": ["Movies"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": { "type": "string" },
              "description": "The movie ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Movie deleted successfully"
            },
            "400": {
              "description": "Invalid ID format"
            },
            "404": {
              "description": "Movie not found"
            }
          }
        }
      }
    }
  }
  